+++ css/new-styles.css
@@ -478,26 +478,27 @@
 
 /* floating  */
 
-.bubble {
+.box {
   background-color: white;
   /* background-image: url(/images/spritesheet.png); */
-  background-size: 1076px 1076px;
-  /* border-radius: 50%; */
+  background-size: cover;
+  /* background-repeat: no-repeat; */
   box-shadow: 0 15px 35px rgba(0, 0, 0, 0.1), 0 3px 10px rgba(0, 0, 0, 0.1);
   height: 82px;
-  margin-right: 20px;
+  /* margin-right: 20px; */
   position: absolute;
   transition: opacity ease-in-out 1s;
   width: 152px;
+  border-radius: 15px;
 }
 
-.bubbles {
+.boxes {
   height: 600px;
   overflow: hidden;
   position: relative;
 }
 
-.bubbles-wrapper {
+.boxes-wrapper {
   /* background-color: salmon; */
   display: flex;
   flex-direction: column;
@@ -564,3 +565,25 @@
 /* .n-stripe.grey2 {
   background: transparent;
 } */
+
+.home-team-container {
+  padding: 50px 0;
+  background: #eaecf0;
+  border-radius: 15px;
+  position: relative;
+}
+
+.home-team-container .n-banner-title {
+  text-align: center;
+  position: absolute;
+  width: 65%;
+  margin: auto;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%) !important;
+}
+
+.client-content {
+  width: 1321px;
+  margin: auto;
+}
