+++ pages/index.js
@@ -8,6 +8,7 @@ import ProjectSlider from "../pageComponents/ProjectSlider";
 import SEOImg from "../images/neo-web-img.jpeg";
 import ServiceSection from "../pageComponents/ServiceSection";
 import Clients from "../pageComponents/Clients";
+import HomeTeam from "../pageComponents/HomeTeam";
 
 const Home = () => {
   const router = useRouter();
@@ -47,6 +48,9 @@ const Home = () => {
           </section>
           <section className="section">
             <Clients />
+          </section>
+          <section className="section">
+            <HomeTeam />
           </section>
         </div>
       </div>
